requirements.txt
transformers==4.27.1
streamlit==1.20.0
pillow>=9.0